Rename drawable namespace to avoid clashes with framework class
